Add an option to skip building API docs when rebuilding web site
(Generally useful, not just in an emergency!)

+++ build/buildweb.sh
@@ -4,19 +4,22 @@ set -e
 
 if [[ "$1" = "" ]]
 then
-  echo "Usage: buildweb.sh output-directory"
-  echo "e.g. buildweb c:\\users\\jon\\NodaTime\\nodatime.org"
+  echo "Usage: buildweb.sh output-directory [--skip-api]"
+  echo "e.g. buildweb.sh c:\\users\\jon\\NodaTime\\nodatime.org"
   echo "It is expected that the output directory already exists and is"
-  echo "set up for git..."
+  echo "set up for git. If --skip-api is set, it is assumed the API docs already exist."
   exit 1
 fi
 
 WEB_DIR="$1"
 
 # Build the API docs with docfx
-./buildapidocs.sh
-rm -rf ../src/NodaTime.Web/docfx
-mv tmp/docfx/_site ../src/NodaTime.Web/docfx
+if [[ "$2" != "--skip-api" ]]
+then
+  ./buildapidocs.sh
+  rm -rf ../src/NodaTime.Web/docfx
+  mv tmp/docfx/_site ../src/NodaTime.Web/docfx
+fi
 
 # Build the web site ASP.NET Core
 rm -rf ../src/NodaTime.Web/bin/Release
